+++ app/src/main/java/lilun/com/pensionlife/ui/tourism/list/TourismFilter.java
@@ -140,13 +140,13 @@ public class TourismFilter {
         }
 
         public static class PriceBean {
-            private List<Integer> between;
+            private List<Double> between;
 
-            public List<Integer> getBetween() {
+            public List<Double> getBetween() {
                 return between;
             }
 
-            public void setBetween(List<Integer> between) {
+            public void setBetween(List<Double> between) {
                 this.between = between;
             }
         }
